+++ ChangeLog.md
@@ -9,6 +9,7 @@
 * Bug fix: Ensure timeout detection can never stop prematurely
 * Fix parsing of datasets with a final SequenceDelimiterItem at the end (#1157)
 * Fix: DicomDirectory did throw exception on calling constructor with no parameters (#1176)
+* Update to DICOM Standard 2021b (#1189)
 
 #### v.4.0.7 (11/1/2020)
 * Bug fix: Not able to open deflated dicom file which contains squence (#1097)